Player wird gerendert. Das PlayerMovement wurde angefangen. Der KeyListener wurde entfernt, stattdessen werden allgemeine Eingaben jetzt direkt verwaltet oder das entsprechende Spielerobjekt kümmert sich selbst um das Movement.

+++ src/jumpignon/Rectangle.java
@@ -10,7 +10,7 @@ public class Rectangle extends RenderItem{
    
     //
 //rotation muss nochmal gedacht werden ;) wegen collisoin
-public Rectangle(int x, int y, int height ,int  width ,int rotation ){
+public Rectangle(float x, float y, int height ,int  width ,int rotation ){
     this.pos_x = x;
     this.pos_y = y;
     this.height = height;
